Code cleanup and refactoring

+++ src/main/java/dev/hilligans/ourcraft/client/rendering/graphics/implementations/splitwindows/FloatingWindow.java
@@ -45,12 +45,12 @@ public class FloatingWindow extends RenderWindow {
     }
 
     @Override
-    public float getWindowWidth() {
+    public int getWindowWidth() {
         return 0;
     }
 
     @Override
-    public float getWindowHeight() {
+    public int getWindowHeight() {
         return 0;
     }
